+++ app/models/import.rb
@@ -13,6 +13,7 @@ class Import < ApplicationRecord
   after_commit :remove_attached_file, on: :destroy
 
   validates :name, presence: true, uniqueness: { scope: :user_id }
+  validate :file_size_within_limit, if: -> { user.trial? }
 
   enum :status, { created: 0, processing: 1, completed: 2, failed: 3 }
 
@@ -57,5 +58,13 @@ class Import < ApplicationRecord
 
   def remove_attached_file
     file.purge_later
+  end
+
+  def file_size_within_limit
+    return unless file.attached?
+
+    if file.blob.byte_size > 11.megabytes
+      errors.add(:file, 'is too large. Trial users can only upload files up to 10MB.')
+    end
   end
 end