+++ spec/requests/api/v1/countries/borders_spec.rb
@@ -10,6 +10,13 @@ RSpec.describe 'Api::V1::Countries::Borders', type: :request do
 
         expect(response).to have_http_status(:unauthorized)
       end
+
+      it 'returns X-Dawarich-Response header' do
+        get '/api/v1/countries/borders'
+
+        expect(response.headers['X-Dawarich-Response']).to eq('Hey, I\'m alive!')
+        expect(response.headers['X-Dawarich-Version']).to eq(APP_VERSION)
+      end
     end
 
     context 'when user is authenticated' do
@@ -21,6 +28,13 @@ RSpec.describe 'Api::V1::Countries::Borders', type: :request do
         expect(response).to have_http_status(:success)
         expect(response.body).to include('AF')
         expect(response.body).to include('ZW')
+      end
+
+      it 'returns X-Dawarich-Response header' do
+        get '/api/v1/countries/borders', headers: { 'Authorization' => "Bearer #{user.api_key}" }
+
+        expect(response.headers['X-Dawarich-Response']).to eq('Hey, I\'m alive and authenticated!')
+        expect(response.headers['X-Dawarich-Version']).to eq(APP_VERSION)
       end
     end
   end